+++ tests/files/wav.rs
@@ -2,8 +2,8 @@ use crate::{set_artist, temp_file, verify_artist};
 use lofty::config::ParseOptions;
 use lofty::file::FileType;
 use lofty::prelude::*;
+use lofty::probe::Probe;
 use lofty::tag::TagType;
-use lofty::Probe;
 
 use std::io::{Seek, Write};